move: source service adds prometheus for /metrics route (#14448)
## Description 

On start, serve Prometheus `/metrics`. Implements metric to count requests.

## Test Plan 

Added a test to check the endpoint--will be useful to double check the
expected format/scheme of metrics as they are added too.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes