[move-compiler] Fix issue with mut borrow warnings. Suppress unused id field (#14545)
## Description 

- Unused mutable borrow warning did not play well with unpack field. The
fields from an unpack now track the original reference, since you cannot
switch individual fields to be immutable
- Suppressed unused field warning for id in Sui mode

## Test Plan 

- New tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

This release fixes two issues with compiler warnings. 
- Fixed an issue where mutable borrow warnings did not work as intended
with unpack patterns. The warning was given when it should not have been
if some fields in the pattern were not used mutably.
- The unused field warning is now suppressed for objects in Sui mode.