BackingPackageStore trait returns Arc (#14470)
## Description 

This PR makes the BackingPackageStore trait ready to support
object/package cache.
BackingPackageStore used to have two APIs, get_package and
get_package_object. Both returns value (MovePackage and Object). This
will make it very inefficient to support a package cache because we
would have to clone all the time.
This PR makes it such that BackingPackageStore trait now only has one
API, which is get_package_object. It returns an Arc of package object
that can be passed around cheaply.
The primary change is in storage/src/mod.rs. All other changes are due
to the API change.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes