Make ConsensusHandler abstract (#14591)
## Description 

This PR completes the `ConsensusOutputAPI` trait, and makes
ConsensusHandler consensus-type agnostic.
There are a few changes that made this possible:
1. The biggest change is the `into_transactions()` function, which hides
the details of the ConsensusOutput data structure and returns all
transactions from it.
2. Introduced the `intro_transactions()` function to `Batch` so that we
can get rid of unnecessary clones. In fact, all previous clones were
removed.
3. Removed the `certificate` field from `SequencedConsensusTransaction`
type, because we don't need it. Instead we only need the authority index
of the origin of the certificate.
4. Added `handle_consensus_output_internal` which operates on common
types only, and hence can be called from both Narwhal and Mysticeti.
5. The only semantic difference is that we no longer report the number
of batches processed metric, which is too specific to Narwhal
implementation. We could add it inside Narwhal if we want to.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes