[benchmark] add option to specify number of shared counters directly (#14544)
## Description 

Right now for our shared counter load, we are only able to specify
`shared_counter_hotness_factor`, which can be confusing to people who
think that `hotness factor = 50` means two shared counters are used. So
this PR adds a way to specify the number of shared counters created/used
directly. This change is backward compatible.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes