[cli] implement Display for ObjectChange and update CLI's output (#14657)
## Description 

Implements the `Display` trait for `ObjectChange` type as part of a
larger effort to have a good-looking output format for the
`SuiTransactionBlockResponse` in the CLI.

Note that the changes in the `write_transaction_response` are temporary
until I get the `Display` trait implemented for each type of
`SuiTransactionBlockResponse`, after which this function will change.

## Test Plan 

<img width="1055" alt="image"
src="https://github.com/MystenLabs/sui/assets/135084671/8f40b60a-3642-4055-a259-7ce0bb32ba74">

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Improves the CLI's formatting output of the `ObjectChange` type when
displaying transaction responses.