More optimizations to package load (#14762)
## Description 

I noticed that we still spend a non-trivial amount of time loading
packages in execution. After digging in, I realized that this is mostly
due to us trying to save all runtime loaded packages.
This PR minimizes that overhead by:
1. Do not write unless it's not already in the runtime loaded package
map.
2. Avoid clone.

This fix is sufficient to make the package load disappear in a local
benchmark, and we are seeing about 10% improvements in throughout in the
baseline mode. So this is good enough for now.
However as a long term fix, I think a proper solution is that we should
introduce a verbose flag to execution. Normally the verbose flag is off.
When we detect a fork, we rerun the execution with the verbose flag on
so that we could collect more information for the debug state dump.

Also fixed the execution-only mode in the local benchmark to use the
package cache, otherwise we would see a ton of unnecessary clones there.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes