Integrate http based downloader with archival, snapshot and sui tool (#14761)
## Description 

Allow using no sign requests for sui tool, archival and snapshot use
cases in sui. User just passes in so sign request as a parameter in
their object store config and right downloader tool is used.
Also cleaned up the abstractions and created new ones so in the future
we can support unauthenticated list requests if need be.