[transfer-to-object] Turn on in testnet, minor updates (#14835)
## Description 

Minor updates and cleanups to execution code related to
transfer-to-object. Nothing about the semantics should be changed with
this.

This also turns transfer-to-object on for testnet and devnet (only).

## Test Plan 

Added an additional test, and made sure existing tests continue to all
pass.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Transfer-to-object functionality is now turned on in testnet and devnet.